Test an invocation whose type javac leaves unsubstituted
Javac substituted every type argument in the invocations in
UnusedInvocationResult.java, so the test never reached the new code in
`InferenceFactory.mentionsUnsubstitutedTypeVariable`.

Rename the file to UnsubstitutedInvocationType.java and give it
invocations whose type javac does leave unsubstituted.  The call to
`unsatisfiableCycle` is issued a spurious "type.arguments.not.inferred"
error if the check is removed; the other such invocations do not
currently produce a diagnostic either way, but they pin down which
invocations the check applies to.  The invocations that javac fully
substitutes remain, as the cases in which the dummy assignment context
is usable as a target type.

Also, an invocation's type can be unsubstituted even when the result is
used, as `new Builder<>(array).self()` shows, so say "no assignment
context" rather than "result is unused" in the documentation.
